Возвращен старый модуль намотки

+++ Winding.h
@@ -14,7 +14,7 @@ struct Winding
 {
   uint16_t turns=0;
   uint8_t step=0;
-  uint8_t speed=30;    
+  uint16_t speed=30;    
   uint8_t layers=0;
   bool dir=1;
 };
